+++ src/lib/bump-stats/types.ts
@@ -1,17 +1,12 @@
-import {
-	JetpackImporter,
-	LocalImporter,
-	PlaygroundImporter,
-	SQLImporter,
-	WpressImporter,
-} from 'src/lib/import-export/import/importers';
-
 export enum StatsGroup {
+	// Studio app
 	STUDIO_APP_LAUNCH = 'studio-app-launch-first',
 	STUDIO_APP_LAUNCH_TOTAL = 'studio-app-launch-total',
 	STUDIO_APP_LAUNCH_UNIQUE = 'local-environment-launch-uniques',
 	STUDIO_IMPORT = 'studio-app-import',
 	STUDIO_EXPORT = 'studio-app-export',
+	// Studio CLI
+	STUDIO_CLI_USAGE_UNIQUE = 'studio-cli-usage-unique',
 }
 
 export enum StatsMetric {
@@ -35,33 +30,3 @@ export enum StatsMetric {
 }
 
 export type AggregateInterval = 'daily' | 'weekly' | 'monthly';
-
-export function getPlatformMetric( platform: typeof process.platform ): StatsMetric {
-	switch ( platform ) {
-		case 'darwin':
-			return StatsMetric.DARWIN;
-		case 'linux':
-			return StatsMetric.LINUX;
-		case 'win32':
-			return StatsMetric.WINDOWS;
-		default:
-			return StatsMetric.UNKNOWN_PLATFORM;
-	}
-}
-
-export function getImporterMetric( importer?: string ): StatsMetric {
-	switch ( importer ) {
-		case JetpackImporter.name:
-			return StatsMetric.JETPACK_IMPORTER;
-		case LocalImporter.name:
-			return StatsMetric.LOCAL_IMPORTER;
-		case SQLImporter.name:
-			return StatsMetric.SQL_IMPORTER;
-		case PlaygroundImporter.name:
-			return StatsMetric.PLAYGROUND_IMPORTER;
-		case WpressImporter.name:
-			return StatsMetric.WPRESS_IMPORTER;
-		default:
-			return StatsMetric.UNKNOWN_IMPORTER;
-	}
-}
